requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.3.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   groq
    #   httpx
    #   openai
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via bs4
blinker==1.7.0
    # via streamlit
bs4==0.0.2
    # via -r cookbook/llms/groq/rag/requirements.in
cachetools==5.3.3
    # via streamlit
certifi==2024.2.2
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via curl-cffi
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   duckduckgo-search
    #   streamlit
    #   typer
curl-cffi==0.6.3
    # via duckduckgo-search
distro==1.9.0
    # via
    #   groq
    #   openai
duckduckgo-search==5.3.0
    # via -r cookbook/llms/groq/rag/requirements.in
exceptiongroup==1.2.1
    # via anyio
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via
    #   phidata
    #   streamlit
groq==0.5.0
    # via -r cookbook/llms/groq/rag/requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   groq
    #   ollama
    #   openai
    #   phidata
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.3
    # via
    #   altair
    #   pydeck
jsonschema==4.21.1
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via
    #   altair
    #   pandas
    #   pgvector
    #   pyarrow
    #   pydeck
    #   streamlit
ollama==0.1.8
    # via -r cookbook/llms/groq/rag/requirements.in
openai==1.23.2
    # via -r cookbook/llms/groq/rag/requirements.in
orjson==3.10.1
    # via duckduckgo-search
packaging==24.0
    # via
    #   altair
    #   streamlit
pandas==2.2.2
    # via
    #   altair
    #   streamlit
pgvector==0.2.5
    # via -r cookbook/llms/groq/rag/requirements.in
phidata==2.4.20
    # via -r cookbook/llms/groq/rag/requirements.in
pillow==10.3.0
    # via streamlit
protobuf==4.25.3
    # via streamlit
psycopg[binary]==3.1.18
    # via -r cookbook/llms/groq/rag/requirements.in
psycopg-binary==3.1.18
    # via psycopg
pyarrow==16.0.0
    # via streamlit
pycparser==2.22
    # via cffi
pydantic==2.7.0
    # via
    #   groq
    #   openai
    #   phidata
    #   pydantic-settings
pydantic-core==2.18.1
    # via pydantic
pydantic-settings==2.2.1
    # via phidata
pydeck==0.8.1b0
    # via streamlit
pygments==2.17.2
    # via rich
pypdf==4.2.0
    # via -r cookbook/llms/groq/rag/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   phidata
    #   pydantic-settings
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via phidata
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via streamlit
rich==13.7.1
    # via
    #   phidata
    #   streamlit
    #   typer
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   groq
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.29
    # via -r cookbook/llms/groq/rag/requirements.in
streamlit==1.33.0
    # via -r cookbook/llms/groq/rag/requirements.in
tenacity==8.2.3
    # via streamlit
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via phidata
toolz==0.12.1
    # via altair
tornado==6.4
    # via streamlit
tqdm==4.66.2
    # via openai
typer==0.12.3
    # via phidata
typing-extensions==4.11.0
    # via
    #   altair
    #   anyio
    #   groq
    #   openai
    #   phidata
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   sqlalchemy
    #   streamlit
    #   typer
tzdata==2024.1
    # via pandas
urllib3==1.26.18
    # via requests
